# backend/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"time"

	"github.com/google/uuid"
)

type Donation struct {
	Uuid          uuid.UUID
	Amount        int64
	Currency      string
	PaymentMethod string
	NonprofitID   string
	DonorID       string
	Status        string
	CreatedAt     time.Time
	UpdatedAt     time.Time
}